requirements.txt
Django>=2.1,<3.0
djangorestframework>=3.9.0,<3.10.0
psycopg2-binary>=2.8,<2.9